misc: qcom: qdsp6v2: use 32-bit variable to store token
Even when we send buffer address with more than 32-bit
to DSP, APR callback handle will return a token with only
32-bit. Hence use a 32-bit variable to store the buffer
address locally, so that comparison between these does
not fail.

+++ drivers/misc/qcom/qdsp6v2/audio_utils_aio.h
@@ -1,6 +1,6 @@
 /* Copyright (C) 2008 Google, Inc.
  * Copyright (C) 2008 HTC Corporation
- * Copyright (c) 2009-2014, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2009-2015, The Linux Foundation. All rights reserved.
  *
  * This software is licensed under the terms of the GNU General Public
  * License version 2, as published by the Free Software Foundation, and
@@ -133,7 +133,7 @@ struct audio_aio_buffer_node {
 	struct list_head list;
 	struct msm_audio_aio_buf buf;
 	unsigned long paddr;
-	unsigned long token;
+	uint32_t token;
 	void            *kvaddr;
 	union meta_data meta_info;
 };
